requirements.txt
Cython
python-libsbml>=5.15.0
libroadrunner>=1.4.18
cobra>=0.6.1
# -e git+https://github.com/opencobra/cobrapy.git@devel#egg=cobrapy
optlang>=1.2.0
# -e git+https://github.com/biosustain/optlang.git@devel#egg=optlang

six
pandas
tabulate
Jinja2
pyexcel
pyexcel-xlsx
pytest
pytest-cov
# bioservices